requirements.txt
torch>=2.5.0
torchvision
numpy<2
opencv-python
timm
scipy
scikit-image
kornia
einops

tqdm
prettytable
tabulate
ipykernel

huggingface-hub>0.25
accelerate


cuda-python==12.6.0
cvcuda-cu12 @ file:////home/dev_samzyu_altostrat_com/cvcuda_cu12-0.15.0-cp310.cp311.cp312.cp313.cp38.cp39-cp310.cp311.cp312.cp313.cp38.cp39-manylinux2014_x86_64.manylinux_2_17_x86_64.whl
opencv-contrib-python
opencv-python-headless
